testing/web-platform/tests/encoding-detection/lv-windows-1257.tentative.html

<!doctype html>
<title>lv windows-1257</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>�is ir rakstz�mju kod��anas tests. Da��s valod�s, kur�s tiek izmantotas lat��u valodas burti, l�muma pie�em�anai mums ir nepiecie�ams vair�k ieguld�juma.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1257", 'Expected windows-1257');
    }, "Check detection result");
      done();
};
</script>
